Create README.md
Architect with approximately 10 years of experience, Expert in making Building and Landscape designs. Has worked on Landscape designs in the Purwakarta area, terminal concepts in the Tangerang city area, houses in the Bandung city area, houses in the East Kalimantan area, houses in the Sukabumi area, designing and supervising public facilities in the oil palm area in East Kalimantan, and designing public facilities and supervising the dam area in the West Java area. Able to make a budget carefully and efficiently so as to minimize expenses.